+++ CHANGELOG.md
@@ -7,6 +7,10 @@ This gem is moving onto its own [Semantic Versioning](https://semver.org/) schem
 Prior to v1.0.0 this gem was versioned based on the `MAJOR`.`MINOR` version of RuboCop. The first release of the
 ezcater_rubocop gem was `v0.49.0`.
 
+## Unreleased
+
+- Add `Ezcater/RailsEnv` cop.
+
 ## v1.0.2
 - Exclude bootsnap cache directory (`tmp/cache`).